2020.14 release
Use SSM Parameters and Secrets for stack generated properties.
Single EFS volume for all instances: shared file system and subdirectory for each IMDB instance.
Move scripts that were in CFTs to scripts on the AMI.
Restore from S3 backup option on central instance creation.
pyramid-AMI-install.sh script to be run on an instance to prep it to generate a Marketplace AMI.